Add Checkstyle and format code
