+++ docs/README.md
@@ -1,4 +1,4 @@
-# Dook User Guide
+# dook User Guide
 
 // Update the title above to match the actual product name
