--- src/main/java/Dook/Commands/Command.java
@@ -1,18 +1,0 @@
-package Dook.Commands;
-
-import Dook.Tasks.TaskList;
-import Dook.Storage.Storage;
-import Dook.Ui.Ui;
-import Dook.DookException;
-
-import java.io.IOException;
-
-public abstract class Command {
-    protected boolean isExit = false;
-
-    public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DookException, IOException;
-
-    public boolean isExit() {
-        return this.isExit;
-    }
-}
