--- src/main/java/Dook/Commands/TaskType.java
@@ -1,3 +1,0 @@
-package Dook.Commands;
-
-public enum TaskType {TODO, EVENT, DEADLINE};
