--- src/main/java/Dook/DookException.java
@@ -1,7 +1,0 @@
-package Dook;
-
-public class DookException extends Exception{
-    public DookException(String message) {
-        super(message);
-    }
-}
