+++ src/main/java/dook/DookException.java
@@ -1,0 +1,7 @@
+package dook;
+
+public class DookException extends Exception {
+    public DookException(String message) {
+        super(message);
+    }
+}
