+++ src/main/java/dook/commands/TaskType.java
@@ -1,0 +1,5 @@
+package dook.commands;
+
+public enum TaskType {
+    TODO, EVENT, DEADLINE
+};
