+++ src/main/java/dook/tasks/Task.java
@@ -1,0 +1,37 @@
+package dook.tasks;
+
+public class Task {
+    protected String description;
+    protected Boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getDescription() {
+        return this.description;
+    }
+
+    public String getStatusIcon() {
+        return isDone ? "X" : " ";
+    }
+
+    public void markAsDone() {
+        this.isDone = true;
+    }
+
+    public void unmark() {
+        this.isDone = false;
+    }
+
+    public String fileFormatted() {
+        return (isDone ? "1" : "0") + " | " + this.description;
+    }
+
+    @Override
+    public String toString() {
+        String status = "[" + this.getStatusIcon() + "]";
+        return status + " " + this.description;
+    }
+}
