--- src/test/java/Dook/Commands/ParserStub.java
@@ -1,7 +1,0 @@
-package Dook.Commands;
-
-public class ParserStub {
-    public ParserStub() {
-
-    }
-}
