--- src/test/java/Dook/Commands/StorageStub.java
@@ -1,9 +1,0 @@
-package Dook.Commands;
-
-import Dook.Storage.Storage;
-
-public class StorageStub extends Storage {
-    public StorageStub(String filePath) {
-        super(filePath);
-    }
-}
