+++ src/test/java/dook/commands/ParserStub.java
@@ -1,0 +1,7 @@
+package dook.commands;
+
+public class ParserStub {
+    public ParserStub() {
+
+    }
+}
