+++ src/test/java/dook/commands/StorageStub.java
@@ -1,0 +1,9 @@
+package dook.commands;
+
+import dook.storage.Storage;
+
+public class StorageStub extends Storage {
+    public StorageStub(String filePath) {
+        super(filePath);
+    }
+}
