+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Dook.Dook
+Hello! I'm dook.dook
 What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
